@@ -11,7 +11,8 @@
     "update": "dt-app update",
     "info": "dt-app info",
     "help": "dt-app help",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "test:api": "jest --config=api/jest.config.js"
   },
   "engines": {
     "node": ">=16.13.0"
@@ -20,26 +21,39 @@
   "license": "ISC",
   "dependencies": {
     "@dynatrace-sdk/app-environment": "1.0.1",
+    "@dynatrace-sdk/app-utils": "^0.4.1",
     "@dynatrace-sdk/client-query": "1.4.3",
     "@dynatrace-sdk/error-handlers": "1.1.1",
     "@dynatrace-sdk/navigation": "1.0.1",
     "@dynatrace-sdk/user-preferences": "1.1.0",
     "@dynatrace/strato-components-preview": "0.104.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-router-dom": "^6.4.1"
+    "@tanstack/react-query": "^5.0.0",
+    "@tanstack/react-query-devtools": "^5.0.3",
+    "immer": "^10.0.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-intl": "^6.5.0",
+    "react-router-dom": "^6.4.1",
+    "simple-zustand-devtools": "^1.1.0",
+    "styled-components": "^6.1.0",
+    "typescript": "^5.2.2",
+    "yup": "^1.3.2",
+    "zustand": "^4.4.4"
   },
   "devDependencies": {
-    "dt-app": "0.107.4",
     "@microsoft/eslint-plugin-sdl": "^0.2.0",
+    "@types/jest": "27.4.0",
     "@types/node": "^16.11.34",
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^5.23.0",
     "@typescript-eslint/parser": "^5.23.0",
+    "dt-app": "0.107.4",
     "eslint": "^8.15.0",
     "eslint-plugin-no-secrets": "^0.8.9",
     "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.5.0"
+    "eslint-plugin-react-hooks": "^4.5.0",
+    "jest": "^27.4.7",
+    "ts-jest": "^27.1.3"
   }
 }
